Beelzebub State Machine AI (#23)
* Added basic AI to Beelzebub

Created a StateMachine AI template to Beelzebub. I suggest we use this pattern for the rest of the enemies, but I wanted to see what y'all think about it. Currently has basic Aggro and Wander states.

* Fixed buggy health + a little refactoring

* Added Idle state and removed Wander

Wander state still exists if we ever want it in the future, but doesn't actually do anything right now.

* Finishing touches